third_party/blink/web_tests/external/wpt/web-animations/animation-model/animation-types/interpolation-per-property-002-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] list-style-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with linear easing
  assert_equals: The value should be url("http://localhost/test-1") at 499ms expected "url(\\"http://localhost/test-1\\")" but got "-webkit-cross-fade(url(\\"http://localhost/test-1\\"), url(\\"http://localhost/test-2\\"), 0.499)"
[FAIL] list-style-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with effect easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "-webkit-cross-fade(url(\\"http://localhost/test-1\\"), url(\\"http://localhost/test-2\\"), 0.428259)"
[FAIL] list-style-image uses discrete animation when animating between "url("http://localhost/test-1")" and "url("http://localhost/test-2")" with keyframe easing
  assert_equals: The value should be url("http://localhost/test-1") at 940ms expected "url(\\"http://localhost/test-1\\")" but got "-webkit-cross-fade(url(\\"http://localhost/test-1\\"), url(\\"http://localhost/test-2\\"), 0.428259)"
[FAIL] word-spacing supports animating as a percentage
  assert_equals: The value should be 30% at 500ms expected "30%" but got "0px"
[FAIL] word-spacing supports animating as combination units "px" and "%"
  assert_equals: The value should be calc(10% + 5px) at 500ms expected "calc(10% + 5px)" but got "5px"
[FAIL] word-spacing supports animating as combination units "%" and "em"
  assert_equals: The value should be calc(5% + 10px) at 500ms expected "calc(5% + 10px)" but got "10px"
[FAIL] word-spacing supports animating as combination units "px" and "calc"
  assert_equals: The value should be calc(10% + 10px) at 500ms expected "calc(10% + 10px)" but got "5px"
[FAIL] word-spacing supports animating as a calc
  assert_equals: The value should be calc(15% + 15px) at 500ms expected "calc(15% + 15px)" but got "0px"
Harness: the test ran to completion.
